Extract Wiktionary lookup helper
