+++ AGENTS.md
@@ -6,17 +6,10 @@
 
 В личных чатах обоих ботов можно отправить сообщение вида `?слово`. Бот
 проверит наличие слова в словаре игры (и добавит его в ответ) и затем
-обратится к ru.wiktionary.org через функцию `lookup_wiktionary`. По
-данным Викисловаря:
-1. если слово отсутствует — сообщает об этом;
-2. если слово найдено, но не является существительным — отвечает, что
-   такого слова нет;
-3. если слово найдено и это существительное — возвращает его первое
-   определение.
-
-Только когда Викисловарь не даёт приемлемого результата, бот направляет
-запрос с этим словом в модель LLM через LangChain и вернёт краткое
-определение, полученное из LLM.
+обратится к модели LLM через LangChain, чтобы получить краткое
+определение. При необходимости для обхода сетевых проблем можно
+использовать вспомогательную функцию `lookup_wiktionary` из модуля
+`wiktionary_utils`, которая обращается к ru.wiktionary.org.
 
 После старта каждой игры бот один раз отправляет всем игрокам сообщение:
 **Новая функция**: в игре можно отправлять знак `?` и слово, чтобы проверить
