+++ tests/test_lookup_wiktionary.py
@@ -7,7 +7,7 @@ import sys
 
 sys.path.append(str(Path(__file__).resolve().parents[1]))
 
-import llm_utils
+import wiktionary_utils
 
 
 class DummyResponse(io.BytesIO):
@@ -29,8 +29,8 @@ def test_lookup_wiktionary_iny():
     def fake_urlopen(url):
         return DummyResponse(payload)
 
-    with patch("llm_utils.request.urlopen", fake_urlopen):
-        exists, is_noun, definition = llm_utils.lookup_wiktionary("инь")
+    with patch("wiktionary_utils.request.urlopen", fake_urlopen):
+        exists, is_noun, definition = wiktionary_utils.lookup_wiktionary("инь")
 
     assert exists is True
     assert is_noun is True
@@ -44,8 +44,8 @@ def test_lookup_wiktionary_missing():
     def fake_urlopen(url):
         return DummyResponse(payload)
 
-    with patch("llm_utils.request.urlopen", fake_urlopen):
-        result = llm_utils.lookup_wiktionary("foo")
+    with patch("wiktionary_utils.request.urlopen", fake_urlopen):
+        result = wiktionary_utils.lookup_wiktionary("foo")
 
     assert result == (False, False, "")
 
@@ -58,8 +58,8 @@ def test_lookup_wiktionary_first_line():
     def fake_urlopen(url):
         return DummyResponse(payload)
 
-    with patch("llm_utils.request.urlopen", fake_urlopen):
-        exists, is_noun, definition = llm_utils.lookup_wiktionary("толь")
+    with patch("wiktionary_utils.request.urlopen", fake_urlopen):
+        exists, is_noun, definition = wiktionary_utils.lookup_wiktionary("толь")
 
     assert exists is True
     assert is_noun is True
